Put kcorrect_spectra in spectrum.py. Changed the last test. Coefficients as argument now. Adapted docstring.
